Separate checkout domain from route registration
