+++ packages/backend/convex/customers/checkout/impl.ts
@@ -1,13 +1,26 @@
+import type { ProductAnalyticsCaptureError } from "@repo/backend/convex/analytics/capture";
+import { isAccountDeletionPending } from "@repo/backend/convex/auth/deletion/state";
 import { getPolarCheckoutLocale } from "@repo/backend/convex/customers/checkout/localization";
 import {
   type CheckoutRequest,
   type CheckoutRequestInput,
+  checkoutSessionIoError,
   InvalidCheckoutSuccessUrl,
   invalidCheckoutSuccessUrlCode,
 } from "@repo/backend/convex/customers/checkout/spec";
 import { products } from "@repo/backend/convex/utils/polar/products";
 import { siteOrigin } from "@repo/backend/convex/utils/site";
 import { Effect } from "effect";
+
+type CheckoutAdmissionUser = Parameters<typeof isAccountDeletionPending>[0];
+
+interface CheckoutAdmissionOperations {
+  readonly captureEvent: () => Effect.Effect<
+    boolean,
+    ProductAnalyticsCaptureError
+  >;
+  readonly loadUser: () => Promise<CheckoutAdmissionUser | null>;
+}
 
 const checkoutProductIds = [products.pro.id] as const;
 
@@ -42,4 +55,30 @@ export const validateCheckoutRequest = Effect.fn(
     productIds: checkoutProductIds,
     successUrl: input.successUrl,
   } satisfies CheckoutRequest;
+});
+
+/** Revalidates account access after Polar IO without gating sales on analytics. */
+export const admitCheckoutProgram = Effect.fn(
+  "customers.checkout.admitCheckout"
+)(function* (operations: CheckoutAdmissionOperations) {
+  const user = yield* Effect.tryPromise({
+    try: operations.loadUser,
+    catch: checkoutSessionIoError,
+  });
+
+  if (!user || isAccountDeletionPending(user)) {
+    return false;
+  }
+
+  yield* operations
+    .captureEvent()
+    .pipe(
+      Effect.catchTag("ProductAnalyticsCaptureError", (error) =>
+        Effect.logWarning("Checkout analytics capture failed.").pipe(
+          Effect.annotateLogs({ code: error.code, reason: error.message })
+        )
+      )
+    );
+
+  return true;
 });
